feat: Create a buffering method that allows clients to store and print messages effectively
In the client side:
Created a function that receives messages with buffering and stores them.

Created a main function that gets the message from the user, prints it and then goes through the list printing the stored messages

Ran them concurrently

In the server side:

Modified the handle_client function to recieve buffered messages and distribute them.